requirements.txt
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
SQLAlchemy-Utils==0.41.1
uvicorn==0.22.0
Werkzeug==2.3.4
python-dotenv==1.0.0
